requirements.txt
streamlit>=1.28.0
streamlit-nested-layout>=0.1.1
streamlit-extras>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
plotly>=5.18.0
statsmodels>=0.14.0
scipy>=1.11.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
openai>=1.3.0
google-generativeai>=0.3.0
jsonschema>=4.20.0
scikit-learn>=1.3.0
statsmodels>=0.14.0